% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agvgd-package.R
\docType{package}
\name{agvgd-package}
\alias{agvgd-package}
\alias{_PACKAGE}
\title{agvgd: An R Implementation of the 'Align-GVGD' Method}
\description{
'Align-GVGD' ('A-GVGD') is a method to predict the impact of 'missense' substitutions based on the properties of amino acid side chains and protein multiple sequence alignments \doi{10.1136/jmg.2005.033878}. 'A-GVGD' is an extension of the original 'Grantham' distance to multiple sequence alignments. This package provides an alternative R implementation to the web version found on \url{http://agvgd.hci.utah.edu/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maialab.org/agvgd/}
  \item \url{https://github.com/maialab/agvgd}
  \item Report bugs at \url{https://github.com/maialab/agvgd/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte.scientist@gmail.com} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
}

}
\keyword{internal}
