% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_pos_to_residue_pos.R
\name{alignment_pos_to_residue_pos}
\alias{alignment_pos_to_residue_pos}
\title{Convert an alignment position to residue position}
\usage{
alignment_pos_to_residue_pos(alignment, alignment_pos)
}
\arguments{
\item{alignment}{An alignment.}

\item{alignment_pos}{An alignment position.}
}
\value{
An integer vector of positions of the residues in the reference
  sequence.
}
\description{
This function converts an alignment position to a position in the frame of
the reference protein sequence, i.e., to the positions of the residues in the
first sequence of the alignment.
}
