% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acids.R
\name{amino_acids}
\alias{amino_acids}
\title{The 20 standard amino acids}
\usage{
amino_acids(code = c("one_letter", "three_letter"))
}
\arguments{
\item{code}{The type of amino acid symbol to be returned, one-letter
(`one_letter`) or three-letter (`three_letter`) codes.}
}
\value{
A character vector of the 20 standard amino acids.
}
\description{
The 20 amino acids that are encoded directly by the codons of the universal
genetic code.
}
\examples{
# By default `amino_acids` returns one-letter symbols
amino_acids()

# Use code = 'three_letter' instead for three-letter symbols
amino_acids(code = 'three_letter')

}
