% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthologs.R
\name{get_orthologue_sequences}
\alias{get_orthologue_sequences}
\title{Get orthologue sequences}
\usage{
get_orthologue_sequences(symbol, format = c("chr", "tbl"))
}
\arguments{
\item{symbol}{A character vector of gene symbols.}

\item{format}{The format of the output. If \code{chr}, a list of named character
vectors is returned. Else, a list of tibbles is returned.}
}
\value{
A list of vectors of orthologue sequences. The list is named with the
names matching the gene symbol. Each element of the list is a character
vector comprising the orthologue sequences of the protein tree for the gene
symbol queried. The first sequence
}
\description{
This function retrieves the sequences of the orthologue proteins in Ensembl's
\href{https://www.ensembl.org/info/genome/compara/homology_method.html}{protein tree} of
the gene indicated by its gene symbol (\code{symbol} argument).
}
\details{
This function communicates with Ensembl's REST API
\href{https://rest.ensembl.org/documentation/info/homology_symbol}{homology/symbol/:species/:symbol}
endpoint.
}
\examples{
\dontrun{
get_orthologue_sequences(c("BRCA1", "BRCA2"))

get_orthologue_sequences(c("BRCA1", "BRCA2"), format = 'tbl')
}
}
