% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_pos_to_residue_pos.R
\name{poi_to_res}
\alias{poi_to_res}
\title{Convert an alignment position to residue position}
\usage{
poi_to_res(alignment, poi)
}
\arguments{
\item{alignment}{An alignment.}

\item{poi}{An alignment position.}
}
\value{
An integer vector of positions of the amino acid residues in the
  reference sequence.
}
\description{
This function converts an alignment position to a position in the frame of
the reference protein sequence, i.e., to the positions of the amino acids in
the first sequence of the alignment.
}
