% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_substitutions.R
\name{read_substitutions}
\alias{read_substitutions}
\title{Read a file with amino acid substitutions}
\usage{
read_substitutions(
  file = stop("`file` must be specified"),
  amino_acid_code = c("one_letter", "three_letter")
)
}
\arguments{
\item{file}{The path to a file with amino acid substitutions.}

\item{amino_acid_code}{The type of symbol used for amino acids in the returned output.}
}
\value{
A tibble listing the amino acids substitutions.
}
\description{
This function reads a file with amino acid substitutions. The format of
should be the same one as requested by the web version of
\href{http://agvgd.hci.utah.edu/help.php}{Align-GVGD}.
}
