% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_pos_to_alignment_pos.R
\name{res_to_poi}
\alias{res_to_poi}
\title{Convert a residue position to an alignment position}
\usage{
res_to_poi(alignment, res)
}
\arguments{
\item{alignment}{An alignment.}

\item{res}{A residue position.}
}
\value{
An integer vector of alignment positions corresponding to residue position in the reference
  sequence.
}
\description{
This function converts an residue position to a position in the frame of
the alignment.
}
