% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_pos_to_alignment_pos.R
\name{residue_pos_to_alignment_pos}
\alias{residue_pos_to_alignment_pos}
\title{Convert a residue position to an alignment position}
\usage{
residue_pos_to_alignment_pos(alignment, residue_pos)
}
\arguments{
\item{alignment}{An alignment.}

\item{residue_pos}{A residue position.}
}
\value{
An integer vector of alignment positions corresponding to residue position in the reference
  sequence.
}
\description{
This function converts an residue position to a position in the frame of
the alignment.
}
