% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chop.R
\name{str_chop}
\alias{str_chop}
\title{Chop a string into sub-strings of fixed width}
\usage{
str_chop(x, width = 50)
}
\arguments{
\item{x}{A single string. This function is not vectorised, so \code{x} must be of
length 1.}

\item{width}{Length of the chopped pieces.}
}
\value{
A character vector of sub-strings of length \code{width}.
}
\description{
Chop a string into sub-strings of fixed width
}
\keyword{internal}
