% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_to_alignment.R
\name{profile_to_alignment}
\alias{profile_to_alignment}
\title{Converts a sequence profile to an alignment}
\usage{
profile_to_alignment(profile)
}
\arguments{
\item{profile}{A sequence profile object as returned by
\code{protean::read_profile()} or \code{protean::get_profile()}. See the \code{{protean}}
package at \url{https://github.com/maialab/protean}.}
}
\value{
An alignment object, i.e. a character matrix whose elements are
protein residues in one-letter notation. Rows are sequences and columns are
alignment positions.
}
\description{
This function converts a sequence profile as provided in the format of the
package \code{{protean}} to an \code{{agvgd}} alignment --- an alignment in this sense
is simply a character matrix whose elements are protein residues in
one-letter notation, rows are sequences and columns correspond to alignment
positions.
}
